+++ CHANGES.md
@@ -9,6 +9,7 @@
   * Feature: Retry writes to storage engines (rather than dying).
   * Feature: Support for snapshot_interval
   * Bugfix: Set arctic storage quota to unlimited (rather than 10G)
+  * Bugfix: Regression in arctic storage backend
 
 ### 0.3.0 (2020-08-18)
   * Feature: Config options for controlling data channel timeouts